# faster permutations, based on what rdeits suggested