--- a/src/permutations.jl
+++ b/src/permutations.jl
@@ -9,14 +9,52 @@ export
     permutations
 
 
-struct Permutations{T}
-    a::T
-    t::Int
+struct PermutationIterator{T}
+    data::T
+    length::Int
 end
 
-Base.eltype(::Type{Permutations{T}}) where {T} = Vector{eltype(T)}
+function has_repeats(state)
+    for outer in 1:length(state)
+        for inner in (outer+1):length(state)
+            if state[outer] == state[inner]
+                return true
+            end
+        end
+    end
+    return false
+end
 
-Base.length(p::Permutations) = (0 <= p.t <= length(p.a)) ? factorial(length(p.a), length(p.a)-p.t) : 0
+function increment!(state, max)
+    state[end] += 1
+    for i in length(state):-1:2
+        if state[i] > max
+            state[i] = 1
+            state[i-1] += 1
+        end
+    end
+end
+
+function next_permutation!(state, max)
+    while true
+        increment!(state, max)
+        if !has_repeats(state)
+            break
+        end
+    end
+end
+
+function Base.iterate(p::PermutationIterator, state=ones(Int, p.length))
+    next_permutation!(state, length(p.data))
+    if state[1] > length(p.data)
+        return nothing
+    end
+    [p.data[i] for i in state], state
+end
+
+Base.length(p::PermutationIterator) = Int(div(factorial(big(length(p.data))), factorial(big(length(p.data) - p.length))))
+Base.eltype(p::PermutationIterator) = Vector{eltype(p.data)}
+
 
 """
     permutations(a)
@@ -25,7 +63,7 @@ Generate all permutations of an indexable object `a` in lexicographic order. Bec
 can be very large, this function returns an iterator object.
 Use `collect(permutations(a))` to get an array of all permutations.
 """
-permutations(a) = Permutations(a, length(a))
+permutations(a) = permutations(a, length(a))
 
 """
     permutations(a, t)
@@ -34,42 +72,52 @@ Generate all size `t` permutations of an indexable object `a`.
 """
 function permutations(a, t::Integer)
     if t < 0
-        t = length(a) + 1
+        return []
+    elseif t > length(a)
+        return []
+    elseif t == 0
+        return [Vector{eltype(a)}()]
     end
-    Permutations(a, t)
+    PermutationIterator(a, t)
 end
 
-function Base.iterate(p::Permutations, s = collect(1:length(p.a)))
+function Base.iterate(p::PermutationIterator, s=collect(1:length(p.a)))
     (!isempty(s) && max(s[1], p.t) > length(p.a) || (isempty(s) && p.t > 0)) && return
-    nextpermutation(p.a, p.t ,s)
+    nextpermutation(p.a, p.t, s)
 end
 
 function nextpermutation(m, t, state)
     perm = [m[state[i]] for i in 1:t]
     n = length(state)
     if t <= 0
-        return(perm, [n+1])
+        return (perm, [n + 1])
     end
     s = copy(state)
     if t < n
         j = t + 1
-        while j <= n &&  s[t] >= s[j]; j+=1; end
+        while j <= n && s[t] >= s[j]
+            j += 1
+        end
     end
     if t < n && j <= n
         s[t], s[j] = s[j], s[t]
     else
         if t < n
-            reverse!(s, t+1)
+            reverse!(s, t + 1)
         end
         i = t - 1
-        while i>=1 && s[i] >= s[i+1]; i -= 1; end
+        while i >= 1 && s[i] >= s[i+1]
+            i -= 1
+        end
         if i > 0
             j = n
-            while j>i && s[i] >= s[j]; j -= 1; end
+            while j > i && s[i] >= s[j]
+                j -= 1
+            end
             s[i], s[j] = s[j], s[i]
-            reverse!(s, i+1)
+            reverse!(s, i + 1)
         else
-            s[1] = n+1
+            s[1] = n + 1
         end
     end
     return (perm, s)
@@ -94,18 +142,18 @@ function Base.length(c::MultiSetPermutations)
     else
         g = [factorial(i) for i in 0:t]
     end
-    p = [g[t+1]; zeros(Float64,t)]
+    p = [g[t+1]; zeros(Float64, t)]
     for i in 1:length(c.f)
         f = c.f[i]
         if i == 1
             for j in 1:min(f, t)
-                p[j+1] = g[t+1]/g[j+1]
+                p[j+1] = g[t+1] / g[j+1]
             end
         else
             for j in t:-1:1
                 q = 0
-                for k in (j+1):-1:max(1,j+1-f)
-                    q += p[k]/g[j+2-k]
+                for k in (j+1):-1:max(1, j + 1 - f)
+                    q += p[k] / g[j+2-k]
                 end
                 p[j+1] = q
             end
@@ -134,7 +182,7 @@ function multiset_permutations(m, f::Vector{<:Integer}, t::Integer)
     MultiSetPermutations(m, f, t, ref)
 end
 
-function Base.iterate(p::MultiSetPermutations, s = p.ref)
+function Base.iterate(p::MultiSetPermutations, s=p.ref)
     (!isempty(s) && max(s[1], p.t) > length(p.ref) || (isempty(s) && p.t > 0)) && return
     nextpermutation(p.m, p.t, s)
 end
@@ -151,7 +199,7 @@ function nthperm!(a::AbstractVector, k::Integer)
     f = factorial(oftype(k, n))
     0 < k <= f || throw(ArgumentError("permutation k must satisfy 0 < k ≤ $f, got $k"))
     k -= 1 # make k 1-indexed
-    for i=1:n-1
+    for i = 1:n-1
         f ÷= n - i + 1
         j = k ÷ f
         k -= j * f
@@ -182,7 +230,7 @@ function nthperm(p::AbstractVector{<:Integer})
     isperm(p) || throw(ArgumentError("argument is not a permutation"))
     k, n = 1, length(p)
     for i = 1:n-1
-        f = factorial(n-i)
+        f = factorial(n - i)
         for j = i+1:n
             k += ifelse(p[j] < p[i], f, 0)
         end
@@ -193,9 +241,9 @@ end
 
 # Parity of permutations
 
-const levicivita_lut = cat([0 0  0;  0 0 1; 0 -1 0],
-                           [0 0 -1;  0 0 0; 1  0 0],
-                           [0 1  0; -1 0 0; 0  0 0]; dims=3)
+const levicivita_lut = cat([0 0 0; 0 0 1; 0 -1 0],
+    [0 0 -1; 0 0 0; 1 0 0],
+    [0 1 0; -1 0 0; 0 0 0]; dims=3)
 
 """
     levicivita(p)