--- a/src/permutations.jl
+++ b/src/permutations.jl
@@ -15,8 +15,8 @@ struct PermutationIterator{T}
 end
 
 function has_repeats(state::Vector{Int})
-    for outer in 1:length(state)
-        for inner in (outer+1):length(state)
+    for outer in eachindex(state)
+        for inner in (outer+1):lastindex(state)
             if state[outer] == state[inner]
                 return true
             end
@@ -27,7 +27,7 @@ end
 
 function increment!(state::Vector{Int}, max::Int)
     state[end] += 1
-    for i in length(state):-1:2
+    for i in reverse(eachindex(state))[begin:end-1]
         if state[i] > max
             state[i] = 1
             state[i-1] += 1
@@ -44,9 +44,9 @@ function next_permutation!(state::Vector{Int}, max::Int)
     end
 end
 
-function Base.iterate(p::PermutationIterator, state::Vector{Int}=ones(Int, p.length))
-    next_permutation!(state, length(p.data))
-    if state[1] > length(p.data)
+function Base.iterate(p::PermutationIterator, state::Vector{Int}=fill(firstindex(p.data), p.length))
+    next_permutation!(state, lastindex(p.data))
+    if state[begin] > lastindex(p.data)
         return nothing
     end
     [p.data[i] for i in state], state
@@ -65,19 +65,12 @@ Use `collect(permutations(a))` to get an array of all permutations.
 """
 permutations(a) = permutations(a, length(a))
 
-"Need special case for strings because we require 1-based indexing, and there is no `size(::String)`"
-permutations(a::AbstractString) = permutations(collect(a), length(a))
-
-"Need special case for strings because we require 1-based indexing, and there is no `size(::String)`"
-permutations(a::AbstractString, t::Integer) = permutations(collect(a), t)
-
 """
     permutations(a, t)
 
 Generate all size `t` permutations of an indexable object `a`.
 """
 function permutations(a, t::Integer)
-    Base.require_one_based_indexing(a)
     if t < 0
         return []
     elseif t > length(a)