--- a/src/permutations.jl
+++ b/src/permutations.jl
@@ -29,7 +29,7 @@ end
 
 function increment!(state::Vector{Int}, min::Int, max::Int)
     state[end] += 1
-    for i in reverse(eachindex(state))[firstindex(state):end-1]
+    for i in lastindex(state):-1:(firstindex(state) - 1)
         if state[i] > max
             state[i] = min
             state[i-1] += 1