--- a/test/Project.toml
+++ b/test/Project.toml
@@ -1,0 +1,2 @@
+[deps]
+OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"