… Project.toml
… test/combinations.jl
… test/multinomials.jl
… test/numbers.jl
… test/partitions.jl
… test/youngdiagrams.jl