I finally learned how to make a pull request, based on this [issue](https://github.com/JuliaMath/Combinatorics.jl/issues/95)

Here I copied in the code that rdeits originally wrote [here](https://discourse.julialang.org/t/slow-permutations/39427/4)

I also updated the tests to use testsets, which provides a slightly nicer user interface when testing.

I also ran the Julia formatter on the files I edited, which made some whitespace changes. Unfortunately github's differ isn't the best at figuring out what whitespace changed, so it might look a little confusing