@natemcintosh 2022-08-08
Hitting an issue with `@inferred` on Julia 1.0. I tried using [juliaup](https://github.com/JuliaLang/juliaup) to download 1.0 on my computer and run it, but couldn't get it to download. Does anyone know of a place I can more run 1.0 to check out what's happening?