@ararslan 2023-01-30
Why change the name of the type from `Permutations` to `PermutationIterator`? The new name is inconsistent with how other types in this package are named, e.g. `Combinations` and `MultiSetCombinations`. Also, though not exported under the old name, other packages may be defining methods for that type.