@thchr 2024-11-07
This made `permutations` type-unstable, returning a `Vector{T}` or `Permutations{T}` depending on the value of `t` in `permutations(a::T, t)`. Was that necessary for the performance gains?

Cf.:

```jl
julia> @code_warntype permutations([1,2,3])
MethodInstance for permutations(::Vector{Int64})
  from permutations(a) @ Combinatorics ~/.julia/dev/Combinatorics/src/permutations.jl:73
Arguments
  #self#::Core.Const(Combinatorics.permutations)
  a::Vector{Int64}
Body::Union{Combinatorics.Permutations{Vector{Int64}}, Vector{Vector{Int64}}}
1 ─ %1 = Combinatorics.permutations::Core.Const(Combinatorics.permutations)
│   %2 = Combinatorics.length::Core.Const(length)
│   %3 = (%2)(a)::Int64
│   %4 = (%1)(a, %3)::Union{Combinatorics.Permutations{Vector{Int64}}, Vector{Vector{Int64}}}
└──      return %4
```