@natemcintosh 2024-11-07
Hi @thchr, it's been a while since I worked on this, so my memory is a little fuzzy. If I remember correctly, it returns `Vector{T}` when `t` was 0 or 1. I _think_ this made it faster, but I don't remember for sure.

It might be worth running some tests to remind us if that was indeed the case.