@thchr 2024-11-07
I think the main problem with that approach - although it may be faster for specific cases of `t=0` and `t=1` - is that it makes the general case type-unstable.

So, if you test "the whole" thing (calling `permutations` and then iterating over it), it becomes slower Julia doesn't know whether to  iterate over a `Vector{T}` or a `Permutations{T}` ahead of time.

E.g., before this PR:
```jl
f(xs) = collect(permutations(xs))
@btime f(1:3); # before:  367.308 ns (20 allocations: 1.30 KiB)
```
And, after:
```jl
julia> @btime f(1:3);
  432.668 ns (17 allocations: 704 bytes)
```
This occurs even though `iterate(::Permutation, ...)` was sped up dramatically by this PR.
I think the best solution is to drop the special-casing of the `t=0` and `t=1` cases. Even though those cases might be sped up, they slow down all other cases.